requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
frozenlist==1.4.1
gevent==23.9.1
geventhttpclient==2.0.2
greenlet==3.0.3
idna==3.6
multidict==6.0.4
numpy==1.24.4
pillow==10.2.0
pkg_resources==0.0.0
python-rapidjson==1.14
six==1.16.0
tritonclient==2.41.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
